feat(quaint): allow wasm32-unknown-unknown compilation; currently fails on native
