+++ Cargo.toml
@@ -68,6 +68,7 @@ features = [
   "pooled",
   "postgresql",
   "sqlite",
+  "connectors",
 ]
 
 [profile.dev.package.backtrace]
