+++ quaint/Cargo.toml
@@ -23,20 +23,28 @@ resolver = "2"
 features = ["docs", "all"]
 
 [features]
-default = []
+default = ["mysql", "postgresql", "mssql", "sqlite"]
 docs = []
 # Expose the underlying database drivers when a connector is enabled. This is a
 # way to access database-specific methods when you need extra control.
 expose-drivers = []
 
-all = ["mssql", "mysql", "pooled", "postgresql", "sqlite"]
+connectors = [
+  "postgresql-connector",
+  "mysql-connector",
+  "mssql-connector",
+  "sqlite-connector",
+]
+
+all = ["connectors", "pooled"]
 
 vendored-openssl = [
   "postgres-native-tls/vendored-openssl",
   "mysql_async/vendored-openssl",
 ]
 
-postgresql = [
+postgresql-connector = [
+  "postgresql",
   "native-tls",
   "tokio-postgres",
   "postgres-types",
@@ -47,11 +55,24 @@ postgresql = [
   "lru-cache",
   "byteorder",
 ]
+postgresql = []
 
-mssql = ["tiberius", "tokio-util", "tokio/time", "tokio/net", "either"]
-mysql = ["mysql_async", "tokio/time", "lru-cache"]
+mssql-connector = [
+  "mssql",
+  "tiberius",
+  "tokio-util",
+  "tokio/time",
+  "tokio/net",
+]
+mssql = []
+
+mysql-connector = ["mysql", "mysql_async", "tokio/time", "lru-cache"]
+mysql = ["chrono/std"]
+
 pooled = ["mobc"]
-sqlite = ["rusqlite", "tokio/sync"]
+sqlite-connector = ["sqlite", "rusqlite", "tokio/sync"]
+sqlite = []
+
 fmt-sql = ["sqlformat"]
 
 [dependencies]
@@ -67,7 +88,7 @@ futures = "0.3"
 url = "2.1"
 hex = "0.4"
 
-either = { version = "1.6", optional = true }
+either = { version = "1.6" }
 base64 = { version = "0.12.3" }
 chrono = { version = "0.4", default-features = false, features = ["serde"] }
 lru-cache = { version = "0.1", optional = true }
@@ -88,7 +109,11 @@ paste = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 quaint-test-macros = { path = "quaint-test-macros" }
 quaint-test-setup = { path = "quaint-test-setup" }
-tokio = { version = "1.0", features = ["rt-multi-thread", "macros", "time"] }
+tokio = { version = "1.0", features = ["macros", "time"] }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies.getrandom]
+version = "0.2"
+features = ["js"]
 
 [dependencies.byteorder]
 default-features = false
